Major edit using Swagger.
Updated the allomorphs/morphemes/glosses trio route to use Swagger. However, code is now more complex, and needs to have either promises implimented or some other way of handling multi-array function pass-back-and-forth. (Not to mention another possible problem: If an utterance is passed in that uses the same word twice, e.g. 'I bought vanilla cookies and chocolate cookies.', the second instance of cookies will be ignored when converted to allo/morph/gloss since it will have been already taken care of the first time.

The entire route perhaps needs to be re-thought out, as far as end goals. Maybe client side is the way to go?